requirements.txt
flask>=1.0.2
numpy>=1.15.4
pandas>=0.23.4
pytz>=2018.7
xlsxwriter>=1.1.2
liquer-framework>=0.2.3
requests>=2.22.0
lxml>=4.4.1
xmljson>=0.2.0